Add class to process all games after game is complete.

+++ db/migrate/20161119143437_add_strength_to_game.rb
@@ -1,5 +1,5 @@
 class AddStrengthToGame < ActiveRecord::Migration
   def change
-    add_column :games, :strength, :decimal
+    add_column :games, :strength, :float
   end
 end
